+++ test/Headroom/Ext/HaskellSpec.hs
@@ -38,7 +38,7 @@ spec = do
   describe "extractVariables" $ do
     it "extracts variables from Haskell source code with Haddock header" $ do
       template <- emptyTemplate @_ @Mustache
-      let comment   = BlockComment "{-|" "-}"
+      let comment   = BlockComment "{-|" "-}" Nothing
           config    = HeaderConfig ["hs"] 0 0 0 0 [] [] comment
           ti        = TemplateInfo config NoExtData Haskell template
           headerPos = Just (1, 13)
@@ -59,7 +59,7 @@ spec = do
 
     it "extracts variables from Haskell source code without Haddock header" $ do
       template <- emptyTemplate @_ @Mustache
-      let comment   = BlockComment "{-|" "-}"
+      let comment   = BlockComment "{-|" "-}" Nothing
           config    = HeaderConfig ["hs"] 0 0 0 0 [] [] comment
           ti        = TemplateInfo config NoExtData Haskell template
           headerPos = Nothing