+++ test/Headroom/Ext/JavaSpec.hs
@@ -32,7 +32,7 @@ spec = do
 
   describe "extractVariables" $ do
     it "extracts variables from Java source code" $ do
-      let comment   = BlockComment "/*" "*/"
+      let comment   = BlockComment "/*" "*/" Nothing
           config    = HeaderConfig ["java"] 0 0 0 0 [] [] comment
           ti        = TemplateInfo config NoExtData Java undefined
           headerPos = Just (0, 2)