+++ test/Headroom/ExtSpec.hs
@@ -28,7 +28,7 @@ spec = do
     it "extracts variables specific for Haskell file type" $ do
       template <- emptyTemplate @_ @Mustache
       let samplesDir = "test-data" </> "code-samples"
-          comment    = LineComment "--"
+          comment    = LineComment "--" Nothing
           config     = HeaderConfig ["hs"] 0 0 0 0 [] [] comment
           ti         = TemplateInfo config NoExtData Haskell template
           expected   = mkVariables
@@ -49,7 +49,7 @@ spec = do
     it "extracts variables specific for Java file type" $ do
       template <- emptyTemplate @_ @Mustache
       let samplesDir = "test-data" </> "code-samples"
-          comment    = BlockComment "/*" "*/"
+          comment    = BlockComment "/*" "*/" Nothing
           config     = HeaderConfig ["java"] 0 0 0 0 [] [] comment
           ti         = TemplateInfo config NoExtData Java template
           expected   = mkVariables [("_java_package_name", "foo")]
@@ -59,7 +59,7 @@ spec = do
     it "extracts variables specific for PureScript file type" $ do
       template <- emptyTemplate @_ @Mustache
       let samplesDir = "test-data" </> "code-samples"
-          comment    = LineComment "--"
+          comment    = LineComment "--" Nothing
           config     = HeaderConfig ["purs"] 0 0 0 0 [] [] comment
           ti         = TemplateInfo config NoExtData PureScript template
           expected   = mkVariables [("_purescript_module_name", "Test")]