+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 - [#57] FIXED: Comment in code block breaks header detection for Haskell/Haddock
 - [#58] BREAKING CHANGE: Extend configuration for defining header margins
 - [#59] Check compatibility of loaded YAML configuration
+- [#62] Sanitize block comment headers with line prefixes
 - Bump _LTS Haskell_ to `16.26`
 
 ## 0.3.2.0 (released 2020-11-26)